Reverted to using default_connector on Aggregate.
I vacilated on this, because I don't like it very much,
but I think it is the winner, because it allows the
entire sql function to remain as a single piece, instead
of being broken up "cleverly" based on extra parameters.

This is both more sensible from an interface standpoint,
and it directly supports using the sql function to make
other decisions.